- New ports: games/ufoai games/ufoai-data
UFO ALIEN INVASION is a strategy game featuring tactical combat against hostile
alien forces which are about to infiltrate earth at this very moment. You are
in command of a small special unit which has been founded to face the alien
strike force. To be successful on the long run, you will also have to have a
research team study the aliens and their technologies in order to learn as much
as possible about their technology, their goals and the aliens themselves.

WWW:    http://ufo.myexp.de/