Scribes is a text editor for GNOME that is simple, slim and sleek, yet powerful.
Scribes focuses on streamlining your workflow.  It does so by ensuring common
and repetitive operations are intelligently automated.  And also by eliminating
factors that prevent you from focusing on your tasks.

The result is a text editor that provides a fluid user experience.  An editor
that is easy and fun to use.  And an editor that ensures the safety of your
documents at all times.

WWW: http://scribes.sourceforge.net/