- add google-gflags-0.1
google-gflags
Commandline flags module for C++

The gflags package contains a library that implements commandline flags
processing. As such it's a replacement for getopt(). It has increased
flexibility, including built-in support for C++ types like string, and the
ability to define flags in the source file in which they're used.

WWW: http://code.google.com/p/google-gflags/