Fix a compiler warning in the AVR-COFF generator that appeared with GCC 4.x, and caused the compilation to fail due to the -Werror option used.
Remove an annoying assembler warning in the ATmega256x patch that was
frequently triggered by large C switch() statements in GCC, and
generally considered to be a regression.  The author of that patch
agreed that the warning isn't particularly useful anyway for a number
of reasons, and recommended to just drop it.