Add new port for linux-kmod-compat - emulation layer to build linux device drivers as FreeBSD kld modules. Details at
	http://info.iet.unipi.it/~luigi/FreeBSD/linux_bsd_kld.html

Two ports for webcam drivers that use this package are coming in
a matter of minutes.

NOTES:
+ This is only for 6.x and 7.x, no support planned for earlier versions.
+ At the moment only i386 but have patches for 64-bit architectures
  that should be committed soon.
+ This port installs a file in /usr/share/mk . After some discussion
  on the ports list this seems, to be the most reasonable thing to do.
  portlint complains about it.
+ portlint also complains because the url in pkg-descr does not contain
  'www'. Can't help it, the url really does not contain 'www'.
+ The emulation code is under active development so expect frequent
  updates to the portversion and related entries.
+ For the same reason, at this time i prefer not to list all individual
  files in pkg-plist. When the thing has settled a bit more, I will reconsider
  this choice.
+ Documentation is a bit lacking - hopefully it is coming too, as part
  of the ongoing development.

And finally - i have almost spent more time to prepare the port
(including discussing choices, location and studying the ports
infrastructure) than writing the linux emulation code. If some of
the ports guru wants to help cleaning up mistakes that i may have
made, i''l be grateful (but please don't set up a static pkg-plist,
for the reasons above).